Added MonadTimes class and instances.
The MonadTimes class provides an operation to compute the cartesian product
of the results of two monadic computations. It can be implemented using >>=
sequentially, but for streams we can also give an interleaved implementation.

If the second argument is a failing computation, the interleaved implementation
fails in finite time even if the first argument may prodice infinitely many results.

Currently no care has been taken to make the interleaving tricky or efficient.
Moreover, the implementation uses fmap and it might be insightful to investigate
the use of liftM instead.

For the stream monad we do not have `liftM id s == s`, e.g.,

    liftM id (Cons 42 Nil)
  = Cons 42 Nil >>= return
  = Single 42 `mplus` Susp (Nil >>= return)
  = Cons 42 (Susp Nil)

The additional suspensions might have an (either good or bad) effect on the
performance of the `mtimes` operation when replacing `fmap` with `liftM`.